+++ changelog.md
@@ -8,10 +8,10 @@
 
 ### New
 
-- Added `detectCountryCode` which does its best to grab the country code from a phone number
+- Added `findCountryCode` which does its best to grab the country code from a phone number
   - This function only looks for **up to** four digit country codes
   - Note: It's important that you provide a formatted number to this function.
-  - Example: `detectCountryCode('+1 444 555 6666') // => '1'` vs `detectCountryCode('+14445556666') // => '1445'`
+  - Example: `findCountryCode('+1 444 555 6666') // => '1'` vs `findCountryCode('+14445556666') // => '1445'`
 - Added `hasCountryCode` which returns a boolean on if a country code is present or not in the phone number
   - Example: `hasCountryCode('+1 555 555 5555') // => true`